backend/requirements.txt
python-telegram-bot
google-generativeai
supabase
python-dotenv